Change to createquote, now only the slash command works (again sorry)

+++ src/main/kotlin/io/github/amerebagatelle/campbotkotlin/Preconditions.kt
@@ -11,6 +11,11 @@ fun logPrecondition() = precondition {
 @Suppress("unused")
 fun messageCommandDeprecation() = precondition {
     if (message?.content?.startsWith("&") == true) {
+        if(message?.content?.contains("createquote") == true) {
+            message?.channel?.createMessage("The command `&createquote` has been removed. Please use `/createquote` instead.")
+            fail()
+        }
+
         val message = message?.channel?.createMessage("This command is deprecated. Please try the slash command instead. (type `/` in the chat bar).  This version of the command only exists for the sake of mobile devices.")
         delay(5000)
         message?.delete()
